requirements.txt
# Document parsing
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.0.0

# NLP and embeddings
sentence-transformers>=2.2.0

# Vector and SQL storage
chromadb>=0.4.0
sqlalchemy>=2.0.0

# Demo
streamlit>=1.28.0

# Language detection (for Arabic)
langdetect>=1.0.9

# Entity extraction for GraphRAG
spacy>=3.7.0

# Clustering for RAPTOR
numpy>=1.24.0
scikit-learn>=1.3.0

# OpenAI for LLM answers
openai>=1.0.0

# Arabic stopwords for GraphRAG
Arabic-Stopwords>=0.4.3